refactored logout widget
